Add initial version of script for zipping packages
For running k8s and docker autotests against this repo
we need to prepare and upload packages first.

This is a POC of script, it should be refactored and improved in
future, but for now we need to have some prototype somewhere, otherwise
it will be hardcoded in jenkins job.

Change-Id: I58c7965b275dd3a86de69f9e341c61edc04cc976